+++ clients/morio/cmd/vars.go
@@ -33,7 +33,7 @@ var clearCmd = &cobra.Command{
 	Long: `Stores an empty string as a new value for a template variable,
 This will always write a custom template variable.`,
 	Example: "  morio vars clear WARP_DRIVE",
-	Args:  cobra.ExactArgs(1),
+	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		SetVar(args[0], "false")
 	},
@@ -46,7 +46,7 @@ var disableCmd = &cobra.Command{
 	Long: `Stores 'false' as a new value for a template variable,
 This will always write a custom template variable.`,
 	Example: "  morio vars disable WARP_DRIVE",
-	Args:  cobra.ExactArgs(1),
+	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		SetVar(args[0], "false")
 	},
@@ -59,7 +59,7 @@ var enableCmd = &cobra.Command{
 	Long: `Stores 'true' as a new value for a template variable,
 This will always write a custom template variable.`,
 	Example: "  morio vars enable WARP_DRIVE",
-	Args:  cobra.ExactArgs(1),
+	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		SetVar(args[0], "true")
 	},
@@ -89,7 +89,7 @@ var getCmd = &cobra.Command{
 If var NAME is not set, this will return an empty string.
 A custom NAME var has precedence over a default NAME var.`,
 	Example: "  morio vars get WARP_DRIVE",
-	Args:  cobra.ExactArgs(1),
+	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		value := GetVar(args[0])
 		fmt.Print(string(value))
@@ -125,6 +125,21 @@ Run 'morio vars export' to see the JSON structure`,
 	},
 }
 
+// morio vars list
+var listCmd = &cobra.Command{
+	Use:     "list",
+	Short:   "List all vars",
+	Long:    "Lists all template variables and their values",
+	Example: "  morio vars list",
+	Run: func(cmd *cobra.Command, args []string) {
+		allVars := GetVars()
+		for key, val := range allVars {
+			fmt.Printf("%s: %v\n", key, val)
+		}
+
+	},
+}
+
 // morio vars rm
 var rmCmd = &cobra.Command{
 	Use:     "rm NAME",
@@ -138,7 +153,7 @@ the default value.
 If you want the variable gone altogether, use 'morio vars clear' to
 set the var to an empty string. Note that you cannot remove default variables,
 but you can override them.`,
-	Args:  cobra.ExactArgs(1),
+	Args: cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		RmVar(args[0])
 	},
@@ -151,7 +166,7 @@ var setCmd = &cobra.Command{
 	Long: `Stores a new value for a template variable,
 This will always write a custom template variable.`,
 	Example: "  morio vars set WARP_DRIVE 9",
-	Args:  cobra.ExactArgs(2),
+	Args:    cobra.ExactArgs(2),
 	Run: func(cmd *cobra.Command, args []string) {
 		SetVar(args[0], args[1])
 	},
@@ -165,6 +180,7 @@ func init() {
 	varsCmd.AddCommand(exportCmd)
 	varsCmd.AddCommand(getCmd)
 	varsCmd.AddCommand(importCmd)
+	varsCmd.AddCommand(listCmd)
 	varsCmd.AddCommand(rmCmd)
 	varsCmd.AddCommand(setCmd)
 }
@@ -209,13 +225,19 @@ func GetVars() map[string]string {
 	for _, file := range defaults {
 		if !file.IsDir() {
 			name := file.Name()
-			found[name] = GetVar(name)
+			// Skip files that start with a .
+			if len(name) > 0 && name[0] != '.' {
+				found[name] = GetVar(name)
+			}
 		}
 	}
 	for _, file := range customs {
 		if !file.IsDir() {
 			name := file.Name()
-			found[name] = GetVar(name)
+			// Skip files that start with a .
+			if len(name) > 0 && name[0] != '.' {
+				found[name] = GetVar(name)
+			}
 		}
 	}