+++ clients/morio/version/main.go
@@ -1,3 +1,4 @@
 // This file is auto-generated
 package version
+
 var Version string = "0.5.4"