+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [client] Use client UUID as Kafa client ID
 - [client] Drop SASL in favor of mTLS
 - [client] Properly handle arrays in client vars
+- [client] When exporting vars, parse them to their correct type
 
 ## [0.5.4] - 2024-11-15